Moved GPU ref impls to cpp files
