+++ dnn-providers/integration-tests/gpu_ref/CMakeLists.txt
@@ -9,6 +9,8 @@ add_subdirectory(kernels)
 
 add_library(
     hipdnn_gpu_ref STATIC
+    src/GpuFpReferenceConvolution.cpp
+    src/GpuFpReferenceValidation.cpp
     src/GpuRefKernelCompiler.cpp
     ${GPU_REF_KERNEL_GENERATED_SOURCES}
 )
